Letterboxd.
Letterboxd can be viewed on the PC  as a desktop app. However I need to improve the installation process before I release it.

So far, the installation process has been done through Google Drive and my gineau pigs  reported some isssues in installing it with issues from the drive etc.